% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data Set}
\format{A dataframe with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from fire station in miles}
  \item{DAMAGE}{X 1000 dollars in damages}
}}
\source{
\url{https://www.crcpress.com}
}
\usage{
fire
}
\description{
A dataset about fire destruction in US homes
}
\keyword{datasets}
