% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Function fora poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
a plot with the distrubution fitting
}
\description{
My Central Limit Function fora poisson
}
