% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Sample Simulation}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{sample population}

\item{iter}{number of iterations}

\item{time}{times}
}
\value{
a plot with random samples
}
\description{
This makes a random sample
}
